Reorganise referral referrer controllers and views
