+++ app/controllers/public_referrals/referrer/check_answers_controller.rb
@@ -1,0 +1,8 @@
+module PublicReferrals
+  module Referrer
+    class CheckAnswersController < Referrals::Referrer::CheckAnswersController
+      def personal_details
+      end
+    end
+  end
+end
