+++ spec/system/public_referrals/user_adds_their_details_spec.rb
@@ -79,7 +79,7 @@ RSpec.feature "Public Referral: About You", type: :system do
 
   def then_i_am_on_the_your_details_your_name_page
     expect(page).to have_current_path(
-      "/public-referrals/#{@referral.id}/referrer-name/edit",
+      "/public-referrals/#{@referral.id}/referrer/name/edit",
       ignore_query: true
     )
     expect(page).to have_title("Your name")
@@ -88,7 +88,7 @@ RSpec.feature "Public Referral: About You", type: :system do
 
   def then_i_am_on_the_your_name_page
     expect(page).to have_current_path(
-      "/public-referrals/#{@referral.id}/referrer-name/edit",
+      "/public-referrals/#{@referral.id}/referrer/name/edit",
       ignore_query: true
     )
     expect(page).to have_title(
@@ -113,7 +113,7 @@ RSpec.feature "Public Referral: About You", type: :system do
 
   def then_i_see_the_referrer_check_your_answers_page
     expect(page).to have_current_path(
-      "/public-referrals/#{@referral.id}/referrer"
+      "/public-referrals/#{@referral.id}/referrer/check-answers/edit"
     )
     expect(page).to have_title(
       "Your details - Refer serious misconduct by a teacher in England"
@@ -128,7 +128,7 @@ RSpec.feature "Public Referral: About You", type: :system do
 
   def then_i_see_the_what_is_your_phone_number_page
     expect(page).to have_current_path(
-      "/public-referrals/#{@referral.id}/referrer-phone/edit"
+      "/public-referrals/#{@referral.id}/referrer/phone/edit"
     )
     expect(page).to have_title(
       "Your phone number - Your details - Refer serious misconduct by a teacher in England"
